+++ src/shared/state.ts
@@ -13,6 +13,7 @@ export function getDefaultState(): DesktopSettingsType {
     credentials: undefined,
     lastAccount: undefined,
     enableAVCalls: false,
+    enableBroadcastLists: true,
     enableChatAuditLog: false,
     enableOnDemandLocationStreaming: false,
     chatViewBgImg: undefined,