+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 - Floating action button in chatlist to start a new chat
+- Broadcast lists as experimental feature
 
 ## Fixed
 - use addr if displayname is not set for webxdc selfName #2803